third_party/blink/web_tests/fast/dom/shadow/shadow-root-new-expected.txt

This test checks the object 'host.attachShadow({mode: 'open'})' and internals.shadowRoot(div) are the same.

PASS root1 is root2
PASS successfullyParsed is true

TEST COMPLETE
